Fix NativeFileExample's signatures

+++ Source/TitaniumKit/examples/NativeFileExample.cpp
@@ -68,7 +68,7 @@ bool NativeFileExample::get_writable() const TITANIUM_NOEXCEPT
 	TITANIUM_LOG_DEBUG("NativeFileExample::get_writable");
 	return false;
 }
-unsigned NativeFileExample::get_size() const TITANIUM_NOEXCEPT
+unsigned long long NativeFileExample::get_size() const TITANIUM_NOEXCEPT
 {
 	TITANIUM_LOG_DEBUG("NativeFileExample::get_size");
 	return 0;
@@ -94,7 +94,7 @@ bool NativeFileExample::createFile() TITANIUM_NOEXCEPT
 	TITANIUM_LOG_DEBUG("NativeFileExample::createFile");
 	return false;
 }
-unsigned NativeFileExample::createTimestamp() TITANIUM_NOEXCEPT
+unsigned long long NativeFileExample::createTimestamp() TITANIUM_NOEXCEPT
 {
 	TITANIUM_LOG_DEBUG("NativeFileExample::createTimestamp");
 	return 0;
@@ -135,7 +135,7 @@ bool NativeFileExample::isFile() TITANIUM_NOEXCEPT
 	TITANIUM_LOG_DEBUG("NativeFileExample::isFile");
 	return false;
 }
-unsigned NativeFileExample::modificationTimestamp() TITANIUM_NOEXCEPT
+unsigned long long NativeFileExample::modificationTimestamp() TITANIUM_NOEXCEPT
 {
 	TITANIUM_LOG_DEBUG("NativeFileExample::modificationTimestamp");
 	return 0;
@@ -165,7 +165,7 @@ std::string NativeFileExample::resolve() TITANIUM_NOEXCEPT
 	TITANIUM_LOG_DEBUG("NativeFileExample::resolve");
 	return "";
 }
-unsigned NativeFileExample::spaceAvailable() TITANIUM_NOEXCEPT
+unsigned long long NativeFileExample::spaceAvailable() TITANIUM_NOEXCEPT
 {
 	TITANIUM_LOG_DEBUG("NativeFileExample::spaceAvailable");
 	return false;
